+++ samples/OpenCVFaceRecognizer.java
@@ -7,8 +7,12 @@ import static org.bytedeco.javacpp.opencv_core.CV_8UC1;
 import static org.bytedeco.javacpp.opencv_face.createFisherFaceRecognizer;
 // import static org.bytedeco.javacpp.opencv_face.createEigenFaceRecognizer;
 // import static org.bytedeco.javacpp.opencv_face.createLBPHFaceRecognizer;
+import static org.bytedeco.javacpp.opencv_imgcodecs.imread;
+import static org.bytedeco.javacpp.opencv_imgcodecs.CV_LOAD_IMAGE_GRAYSCALE;
 
 import org.bytedeco.javacpp.BytePointer;
+import org.bytedeco.javacpp.IntPointer;
+import org.bytedeco.javacpp.DoublePointer;
 import org.bytedeco.javacpp.opencv_face.FaceRecognizer;
 import org.bytedeco.javacpp.opencv_core.Mat;
 import org.bytedeco.javacpp.opencv_core.MatVector;
@@ -79,7 +83,10 @@ public class OpenCVFaceRecognizer {
 
         faceRecognizer.train(images, labels);
 
-        int predictedLabel = faceRecognizer.predict(testImage);
+        IntPointer label = new IntPointer(1);
+        DoublePointer confidence = new DoublePointer(1);
+        faceRecognizer.predict(testImage, label, confidence);
+        int predictedLabel = label.get(0);
 
         System.out.println("Predicted label: " + predictedLabel);
     }